Improve links consistency
Now when talking about the `Actor` class, a link to the class is used
more consistently. When mentioning the `_run()` method, a link to the
"The `_run()` Method" section is provided too. When mentioning frequenz
channels, we link using the module instead of using URL. This improves
consistency but also makes sure we link to the version the SDK is
currently using.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>