Add a section explaining how to initialize an actor
The initializer is not trivial, as it is usually the one taking input
and output channels, and the `Actor.__init__()` needs to be called too,
and the optional `name` should be explained too.

The new section includes an example and a continuation of this example
is added to the "The `_run()` Method" section, to make it more clear
what the bare minimum needed to implement a complete actor.

We also add the `name` argument to the complete examples.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>