Auto-increment port number in microgrid client tests
This is a hack to avoid the inherent flakiness of the approach of using
a real GRPC (mock) server. The server seems to stay alive for a short
time after the test is finished, which causes the next test to fail
because the port is already in use.

This is a workaround until we have can get rid of the real GRPC server.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>