Update pydantic to v2.3
This update is necessary to stay current with the latest pydantic
version, rather than pinning the library to v1.9.0. However, it's
worth noting that this is a temporary solution until Frequenz-SDK
is updated to use Marshmallow instead of pydantic, which would
require a higher level of effort.

One important consideration is the dependency on
typing-extensions>=4.6.1 introduced in pydantic v2.0.

Additionally, there's a breaking change in pydantic v2,
which disallows data type conversion if it results in a loss
of precision. This is particularly relevant when converting a
float to an int. Consequently, the `test_config.py` file has
been updated to accommodate this change.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>