Expose fill_value option in moving window's window method
The `fill_value` option imputes missing values `NaN` or `None` in the
moving window with the corresponding value. If `None` is passed, no
imputation is done. By default, missing values are imputed with `NaN`.
This prevents returning outdated return values accidentally.

If missing values should be filled, the force_copy argument has to
be true to avoid overwriting the underlying data.

Signed-off-by: cwasicki <126617870+cwasicki@users.noreply.github.com>