Use `-sys.maxsize - 1` as default priority for BatteryPool
When no priorities are specified, we should use the lowest possible
priority.  Zero was supposed to be this lowest possible value, until
it was mentioned that negative priorities should be possible as well.

This commit also updates the BatteryPool's `charge` and `discharge` to
use the pool's priority rather than the minimum.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>