Remove support for `include_broken_batteries` in control commands
The following items are removed:

- the `include_broken_batteries` fields in request objects and from
  function parameter lists,
- the caching of battery data, that was necessary when trying to use
  broken batteries,
- corresponding tests

This was supposed to be for handling battery data outages, but where
the batteries themselves are connected and usable. But this has
stopped being a issue, because of more stable bms firmware.

This also interferes with the safety features on the microgrid
service. So this can be removed.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>