Expose `ComponentGraph` from its own module
The `ComponentGraph` is a (relatively) low-level interface, but was
exposed as a member of the `frequenz.sdk.microgrid` package, and was
available from the documentation of the `microgrid` package.

Moving it to its own package cleans up the docs page, showing just the
high level functions like `initialize`, `logical_meter`,
`battery_pool`, etc. as members of the `microgrid` package.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>