Rename datetime to timestamp variables
There were internal variables in OrderedRingBuffer and MovingWindow
that were named with the prefix `datetime` as a way to refer to the
timestamps that are stored, which happened to be represented using
a datetime object, but timestamps are actually what it is stored.
To avoid confusion when reading the code, these variables were
renamed from `_datetime` to `_timestamp`.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>