Change internal presense of the Grid, eliminating _grid
We are moving the grid to a new location to improve documentation
visibility of the type `Grid`.

This is not particularly easy, because we noticed that the
place where the grid module would fit the best (microgrid.grid),
would cause name conflicts. We agreed to move it to timeseries for now,
which is actually consistent to the current state, and move it further
later.

Signed-off-by: Christian Parpart <christian@parpart.family>