Initialize grid in data pipeline
The grid is initialized in the data pipeline, so that it can
instantiate the formula engine pool to generate grid metrics.
This caused circular imports, which is resolved by enclosing
the type hints in a `TYPE_CHECKING` block.

Also the component metadata is now retrieved from the component
graph, instead of the microgrid API client. This is because
asynchronous calls to the API are not suitable in the data
pipeline as this is a high level user facing interface.

The grid initialize() function needed to be changed to process
the metadata as a dictionary rather than as a ComponentMetadata,
and this is due to the fact that the component graph does not
effectively retrieve the metadata as a ComponentMetadata object.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>