Allow `_clamp_to_bounds` to clamp power values to both sides
In cases where there's an exclusion zone within the usable bounds, the
given power can be clamped to both ends of the exclusion zone.

This function now returns both sides, rather than just the one that's
closest to the given power.

This can also be used in the bounds interface, exposed through the
`Report` class.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>