Refactor Grid interface
Previously, the Grid interface returned `None` if the microgrid was
configured as an island.
This update ensures that the interface consistently returns a Grid
instance.
This change is necessary to treat the grid connection uniformly within
the microgrid, regardless of whether the microgrid is configured as
an island or not. Moreover, it guarantees that the `microgrid.grid()`
function always yields a Grid instance for metric retrieval.

Now the rated current of the fuse is set to zero in case of an
islanded microgrid.
And the fuse of the grid connection point is now set to `None` when
the grid connection component metadata lacks information about the
fuse. Thus, a fuse set to `None` won't prevent the initialization
of the Grid instance.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>