Take bounds from the correct component pool in PowerManagingActor
This is in preparation for using the PowerManagingActor with different
component pools.

The power manager needs to get system bounds for the given set of
components from the right place, and only when necessary.  This
requires telling the power manager which pool to get bounds from, so
that it can do it dynamically.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>