WIP: Force explicit casts when using the ChannelRegistry
This commit creates a new type `Unknown` to be used as the opposite of
`typing.Any`, i.e. to indicate that the type is unknown and forcing the
user to explicitly cast from `Unkown` to a known type.

The problem with this approach is it is not a pleasant interface to work
with, and requires a lot of casting, specially in the tests.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>