Find first descendant component in component graph
Add a generic function to find the first descendant
source component in the component graph given the root
category to find the root component and a list of categories
to find the first descendant component based on the priority
of the categories.
The function can then be used to fetch frequency or voltage
metrics from the grid source component.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>