Fix missing conversion to `QuantityT`
The underlying channel is transporting `Quantity` objects, but the
`ResampledFormulaBuilder` is returning `QuantityT`, i.e. a specific
sub-type of `Quantity` (which has more methods than `Quantity`).

If this object is used to access the extra methods, then it will fail.

This bug was never triggered because the actual value is just always
used as an abstract `Quantity` (only `.base_value` is used), so the
conversion is not needed strictly speaking, but we need to use the type
system properly, and fix the issue at another level.

This will be properly fixed when we make `Quantity` abstract, and we
allow using a raw `float` for stuff that can be parametrized with
quantities. In this case the resampler will just send raw `float`s, so
we will not need to use `Quantity` at all when we are just operating
with raw data, and only apply the finally concrete `Quantity` type when
the data reaches the end user.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>