Speedup tests by lowering the actors restart delay
A restart delay was added by #741 and as a side effect the tests became
almost 40% slower in total and the actor test became 28 times slower
(locally this PR added around 27 seconds to the test run).

This commit fixes this by adding a test fixture to automatically reduce
the restart delay to 0 for tests. It also adds a context manager to
temporarily change the restart delay in case a different value is
needed.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>