Fix checks in tests
The checks in the data sourcing actor were meant to check if the value
of the sample was `None`, not the sample itself (as the type system
guarantees that `sample` will never be `None`.

This issue wasn't detected because the `ChannelRegistry` used `Any` as
message type, so `sample` had type `Any`, which means that as far as
the type system is concerned, it could have been `None`.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>