MockMigrogrid/Resampler: Fail on unhandled exceptions
The tasks created by the MockMigrogrid/Resampler were not handling
exceptions, so if anything goes wrong when streaming (or when setting up
the streaming), the task will silently die, which means no values are
sent to the channels, so tests will probably hang forever waiting for
values.

We now make sure that tasks are cleaned up from the lists when they are
done and also to exit the tests if there was an unhandled exception, so
at least the test will fail and show the origin of the raised exception.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>