Improve finalization of grid tests
Use an async exit stack to properly finalize the `Grid` object
automatically. The finalization was missing in some tests.

Also uses `time_machine` as a `pytest` fixture when possible to simplify
the tests and avoid having too indent too deep, as the previous approach
used a context manager too, but not an async one, so it can be combined
in the same `with` statement.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>