Don't use `Sample` in the resampler buffer and function
`Sample` can have `None` as value, but the resampler buffer (and
function) never get any `None` value stored/passed. The reason is saving
a sample with no value provides no extra information and would make the
resampling function implementation more complicated, as it would need
to check values for `None`.

Currently, the resampling function will never receive a `None` value, but
it still needs to check for them for type-checking reasons.

This commit uses a `tuple[datetime, Quantity]` to store samples in the
buffer and pass it to the resampling function. This way it is trivially
clear that values can't be `None` in this context.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>