Improve error handling when background services are not initialized
Often during the early stages of development, people might just try to
run pytest, and not pylint every time.

This makes it hard to identify when they forget to call
`super().__init__()` when deriving from `BackgroundService`.

This commit improves error handling in `BackgroundService` to provide
a more meaningful error and a solution to an otherwise hard to debug
issue.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>