Retry only once on PartialFailure
When the PowerManager receives a `PartialFailure`, in many cases,
resending the request would fix it, because the power distributor
would have new information on how to redistribute differently.

But this doesn't always work, for example in single battery locations
or if all batteries are in an UNCERTAIN state, in which case the power
distributor has to retry with the same batteries.

For this reason, if we don't have a limit on such retries from the
PowerManager, we'd get into an infinite loop that doesn't have any
delay inbetween retries.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>