Expose 3-phase power through the grid
The 3-phase power is needed to calculate the power factor
of the microgrid. The power factor is not available through
the microgrid API in the current version used in the SDK.

So the 3-phase power is temporary exposed through the grid
until the migration to the latest version is completed.
Then the 3-phase power factor can be fetched and streamed
through the microgrid.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>