Make sure `Quantity.base_value` is always a float
In Python `float` is a bit weird, because `isinstance(1, float)` is
`False` but still `int` is a subtype of `float` in terms of the type
system, so `int` will be accepted when a `float` is specified in
arguments for example.

This leads to quantities that can have `int` as base value if it was
constructed with a `int` instead of a `float`, which is unintuitive
when the `base_value` type hint is `float` and can potentially cause
problems.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>