Allow all quantities multiplication by float | Percentage
This is only applying a factor so it should be safe for any Quantity.

To be able to use nicer names in overloads, we use positional-only
arguments for `__mul__()`, which in practice shouldn't be much of a
breaking change as these methods should be used mostly via operators.

We also use `match` syntax in the changed methods.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>