Make `Quantity` more directly not default constructible
We are going to make `Quantity` a abstract base class, so we don't need
to do the trick to only allow to call the default constructor for the
`Quantity` class but not sub-classes. We just make `Quantity.__init__()`
always raise a `TypeError`, as subclasses can use `_new()` to create
instances.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>