Allow all quantities division by float | Self
Dividing by `float` is only applying a factor so it should be safe for
any `Quantity`.

Dividing by `Self` produces a unit-less ration (`float`), so it should
be safe too.

To be able to use nicer names in overloads, we use positional-only
arguments for `__truediv__()`, which in practice shouldn't be much of a
breaking change as these methods should be used mostly via operators.

We also use `match` syntax in the changed methods.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>