Make quantities convertible to `float`
This was possible before too, but using the `base_value` property, which
is removed by this commit. Now the more idiomatic `float(quantity)`
conversion is provided, although it should only be used for generics,
when functions or classes are parametrized with a `SupportsFloat` type.

`float(quantity)` always returns the quantity in the `base_unit`.

This allows to make generic classes and functions that can operate on
both `Quantity` or raw `float`s (or any other type that supports float).

The methods `isnan()` and `isinf()` are removed because now
`math.isnan()` and `math.isinf()` (as well as other `math` utility
functions) can be directly used with `Quantity`.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>