Parametrize the resampler with `SupportsFloatT`
The resampler can now be paremetrized on both quantities and `float`.
Even more, the values a resampler resamples are now attached to the
specific type, not to the base `Quantity` class, so users can be sure
the resampler will always return the appropriate quantity.

If a resampler must be used to resample different types of quantities,
it can now be parametrized using plain `float`s, which should also yield
better performance as it avoids creating `Quantity` objects.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>