Add a new `ResamplerActorConfig` class
This is just a subclass of `ResamplerConfig` that is specialized with
`float`. Since the `ResamplerConfig` now needs more configuration
(the type and the value constructor), and the `ResamplerConfig` is
passed around quite a bit during initialization, it makes sense to have
a class that already provides the basic stuff for `float`, which is what
will always be used by the resampling actor because it needs to deal
with different types of quantities (and is also good for performance
reasons).

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>