Amend documentation for a method in actor class
The previous documentation of the method `_run_loop()`
was outdated and did not accurately describe the
intention and behavior of the method.

This update removes mention of the restart limit as
it was primarily introduced for testing purposes and
the actor should be always restarted in case of an
unhandled exception.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>