Use `abc.Set` for component_ids in EVChargerPool
`abc.Set` allows us to treat `set`s as read-only, and interchangably
with `frozenset`s.  Many of the SDK's methods that give out component
IDs produce `abc.Set` values, and with this change, such values can
directly be used in calls to `ev_charger_pool`.

This is also to be consistent with `battery_pool`, which already
accepts `abc.Set`.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>