Don't pass blindly what comes from the API to `Fuse`
Since now we have an internal `Fuse` type that uses quantities to avoid
bugs, we need to make sure to properly convert between the client's
`Fuse` type and the internal one.

There is also actually a bug in the old code, as the conversion to
`Current` never happened. This was never caught by `mypy` because we
were using the `**args` syntax to create the object, so there is no
type checking of the arguments, so a bad `Fuse` was being created, with
a `float` `max_current` instead of using `Current`.

We now make sure the conversion is done appropriately.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>