Increase BatteryStatusTracker test tolerance for missing data
Because we were waiting 0.1 seconds to confirm there's no status
change, we were getting into a race, because the 0.1s timeout
finishes, the status changes because of missing data.

Increasing the tolerance for missing data fixes that.

This commit also increases the timeout for failures due to stale data,
to be clearly higher than `max_data_age`.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>