Send power distribution results from the component manager
Until now, the PowerDistributingActor has supported only distributing
power to batteries.  Batteries are (supposed to be) always available,
so the distribution takes place once when the power distributor
receives a request.

This changes when distributing power to EV chargers, because power
needs to be:

- distributed when an EV is connected
- redistributed to other EVs when an EV is disconnected
- redistributed to other EVs if an EV appears to be not charging, etc.

So a power distribution manager for EVs would be best implemented with
its own `select` loop, where distribution happens.

Because of this, a call to the `distribute_power` method on the EV
charger manager can't immediately return a single distribution result,
but that call is not the only thing that triggers the distribution,
but other events too.

So the manager needs to be what sends out the distribution results.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>