Implement a PV pool with a `_system_power_bounds` method
These system bounds trackers need to be moved out of the `*Pool`s, but
exactly how, is yet to be decided.  Until then, I'm following the same
pattern here as in the other `*Pool`s.

This commit adds a `_system_power_bounds` method only because that's
all that's needed for the PowerManager to support PV inverters.

Other methods will be added later.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>