Start EV charger manager only if there are EV chargers in the graph
The EV charger manager has its own task that tracks the status of the
EV chargers and redistributes power to them as EVs are connected and
disconnected.  This approach is specific to the EV charger pool,
because the status of the components is dynamic, unlike the
BatteryPool and the PVPool.

This task needs to run only in locations where there are EV chargers,
because it expects that there be at least one EV charger in a location
and runs into issues if that's not the case.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>