Ignore reactive power in test
The test is incomplete as reactive power is not really used or verified.

Fixing this takes quite a bit of time, as we need to patch the
`MockMicrogrid` which is taking more time than expected, so for now we
are just ignoring the return value here to avoid a `pylint` CI failure
preventing the release.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>