Update release notes for rc6.post1
Semver is not fully supported by setuptools_scm, so we need to use
a version number that is compatible with PEP 440. This means intead of
using `rc6.1` we need to use `rc6.post1`. Using pre-release versions
with pre-release versions is discouraged but in this case it seems
better than do the jump to `rc7`.

Because `rc6.1` (or `rc6.2`) was ever released in PyPI, we'll use
`rc6.post1`.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>