Set zero power for PV inverters not neccessary to reach target power
If desired power is already reached without the need to use some
inverters, their powers need to be set to zero.

If not, their original powers would remain and the target powers would
be wrong.

In case the requested power is zero, this was also leading to crashes
because there were no allocations.  This change also fixes that issue.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>