Use a server URL to connect to the microgrid API
Instead of using host and port, we now use a URL. The following format
is expected: `grpc://hostname{:port}{?ssl=ssl}`, where the port should
be an int between `0` and `65535` (defaulting to `9090`) and `ssl`
should be a boolean (defaulting to `false`). For example:
`grpc://localhost:1090?ssl=true`.

This also removes the default host and port, as normally users always
want to override it.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>