Stop using the `mock_api` for testing
The `mock_api` is using a real gRPC server, which is complicated and
flaky. Instead of using a full blown gRPC server we now use mocks to
mock the client directly.

In the future, the microgrid client will provide a fake client to make
it much more easy for users to test their code. This commit focuses on
removing the use of `mock_api` preserving the test coverage as much as
possible, and *not* on having a nice fake microgrid client in the SDK.
This will come in the future.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>