Update conditions for CI *-all jobs based on results
The current condition for the `*-all` jobs in the CI were
not correct, as it was skipping the job if the matrix job
succeeded, which is not the desired behavior. We want to
skip the *-all job only if the matrix job was skipped.
The results for `*-all` jobs are now set based on the
matrix job results.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>