Handle proposals from power shifting actors in the PowerManager
This is done by having an additional `Matryoshka` instance for
resolving requests from shifting actors.  And the target power thus
obtained is added to the target power from the regular actors, which
in effect, shifts the target power of the regular actors by the target
power of the shifting actors.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>