Update ConfigManagingActor to send a dictionary
The ConfigManagingActor now sends a dictionary instead of the
Config object. The validation code done through the Config
class was ancient without any real purpose.
Sending a dictionary is more flexible and allows the user to
use the configuration as they see fit. The user can then
validate the configuration using marshmallow, pydantic, or
any other library. Later, the frequenz-SDK will provide
functionality to validate configuration based on the schema
provided by the user, still needs to be assessed if it will
be done with pydantic, marshmallow, or another library.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>