Update PowerDistributor tests to not use the wait flag anymore
And instead sleep for the same amount of time, before sending requests
to the power distributor.

This sleep is needed only in the direct tests of the PowerDistributor.

In the `*_pool_control_methods` tests, PowerDistributor receives
requests from the PowerManager, which would happen only when there are
data, so additional sleep is not required in those tests.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>