Use a more descriptive channel name for the resampler
When trying to debug issues with the resampler not getting relevant
data, the name of the channel is being used to identify the component
and metric with issues, but the string contains other data, which makes
it hard to identify the component and metric in the string.

This commit uses a new, more descriptive channel name for for any
component metric request, which is also used by the resampler.

The previous string was:
`component-stream::{component_id}::{metric_id.name}::{start_time}::{namespace}`

The new string is:
`component_metric_request<namespace={namespace},component_id={component_id},metric_id={metric_id.name},start={start_time}>`

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>