Update sender type in ConfigManagingActor
The ConfigManagingActor now sends a collections.abc.Mapping
type as the output sender type to indicate to the user that
the broadcasted configuration is intended to be read-only.
This change leverages type checkers (e.g., mypy, etc.) to
help ensure that users do not attempt to modify the
dictionary contents.

Note that this change won't prevent users from modifying the
contents at runtime, but it will help prevent accidental
modifications.

Signed-off-by: Daniel Zullo <daniel.zullo@frequenz.com>