Remove dependence on `_ComposableFormulaEngine` in `FormulaEngine3Phase`
`_ComposableFormulaEngine` is a generic class exposing arithmetic
methods with typing for single-phase and three-phase formulas.  But
typing support has been incorrect and incomplete since the beginning.

The cleanest solution appears to be to just implement the methods on
each of the `FormulaEngine`s separately, so that they can be fully
typed.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>