Catch `CancelledError`s properly
The method `exception()` in `Task` doesn't return `CancelledError` if
the task was cancelled, instead it raises the error, so if a task is
cancelled an exception would be raised unintentionally.

Instead we want to catch `CancelledError` and treat it like any other
task error for these tasks, are they are not supposed to be cancelled.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>