Avoid unnecessary restarts of the resampling actor
When the subscriptions task ends, the actor is raising a
`RuntimeException` which causes the actor to restart, which means
re-creating not only the subscriptions task but also the resampling
task, which could introduce glitches into the resampling streams
unnecessarily.

In this commit we just restart the subscriptions tasks instead of the
whole actor.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>