Rewrite CI: use reusable actions, native arm and faster PR testing
This commit rewrites the CI configuration to use reusable actions for
testing with nox and setting up Python.

It also uses the new native arm runners, upgrade to use the Ubuntu 24.04
GitHub runner and split the CI into two workflows: one for PRs and one
for pushes.

The PR workflow only runs the nox tests for the default Python version
and test the documentation website generation, so it should be much
quicker.

When pushing (to any branch, including merge queues, and tags), a more
through CI is run, including all the nox sessions for all supported
Python versions, archs and OSes, building the distribution packages,
testing the installation of the built packages, publishing the
documentation website, creating a GitHub release and publishing the
packages to PyPI.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>