Use a custom formula in the higher order formula test
When testing the `FormulaFormatter`, the `microgrid.pv_pool().power` and
`microgrid.grid.power` formulas were used. This pulls a lot of
complexity in, as the whole microgrid needs to be mocked. It also makes
the test nondeterministic, as the components returned by the component
graph, which are used to build the formulas, are returned as a set,
which doesn't provide order, so the resulting formula string will vary
from run to run.

To test higher order formulas, we now build custom formulas instead.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>