Improve `test_producer` output
Change the name from "test case" to "test step" in the fallback test,
as they are not really isolated test cases, but steps in one long test
(as next steps depend on the previous step to have happened). This makes
it more clear that steps are connected and need to run in sequence.

Also the index was made part of the test steps definition, so we can
show more clearly which (sub-)stop failed, and we also add a print line
indicating which step is being ran.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>