Don't rely on the order of a battery set in tests
These tests rely on the order in which components are returned from the
component graph using `all_batteries[:2]` to get the first 2 batteries.

But the component graph returns a `set`, which doesn't have any ordering
guarantees, so if the Python implementation changes, the hashing can
change and this resulting order could be different, breaking the tests.

This is why we are now specifying IDs to use explicitly (these are the
IDs that were used when the tests were written). This can also change in
the future if generator of mock components changes, as it might produce
different IDs, so this solution is also not bullet-proof.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>