Sort component IDs in exceptions
These component IDs are stored in sets, which don't have a order
guarantee, and looking at component IDs in exceptions with random order
makes it hard to identify if a particular component is in the list or
not.

This also makes it easier to write tests that are not flaky or depend
on the component's order.

To improve the exceptions also `", ".join()` is used to get less noise
in the exception message (like `frozenset({4, 2, 6})` vs `2, 4, 5`).
This makes exceptions easier to read because it is easier to look for
a component ID

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>