Treat warnings as errors in tests
We need to make an exception for `PytestUnraisableExceptionWarning`
because we have some cleanup issues, and sometimes exceptions are raised
in `__del__` methods, for example when trying to do async stuff and the
event loop is already closed. This could even be caused by the GC, some
time after the test function that has the issue was completed.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>