Update tests to support latest time-machine
And migrate to latest time-machine version v2.16.0

This new version of `time-machine` mocks the monotonic clock of
asyncio.

This was causing the
`test_battery_status.py::TestBatteryStatus::test_async_battery_status`
test to fail, because `channels.Timer`s weren't working properly.

To fix this, this commit switches this test file to use an
`async_solipsism` event loop instead, which `time-machine` doesn't
affect.

But introducing the `async_solipsism` event loop caused the
`TestBatteryStatusRecovery::test_stale_data` test to fail, because it
needed the clock time to make progress, which wasn't happening because
`sleep` was returning immediately.

To fix this, this commit also updates that test to use `time-machine`
and moves the time forward manually, so the test can pass.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>